requirements.txt
python-dotenv
langchain
langgraph
langchain-community
langchain-google-genai
langchain-chroma
jq
confluent-kafka
PyYAML
apache-airflow
